Optimize bulk data ingestion for dictionaries (#191)
* Use unzip and unnest to bulk insert DB rows

Mostly speeds up inserting dictionary entries, which is the longest
section of our ingestion process.

* Condense all DF1975 inserts into one transaction

* Add sqlx data type info

* Fix DF1975 adjective and verb ingestion

+++ types/src/form.rs
@@ -122,7 +122,7 @@ impl AnnotatedForm {
     }
 
     /// Number of words preceding this one in the containing document
-    async fn index(&self) -> i32 {
+    async fn index(&self) -> i64 {
         self.position.index
     }
